optimization

+++ Service/HotelFeatures/Commands/DeleteHotelByIdCommand.cs
@@ -24,12 +24,17 @@ namespace Service.HotelFeatures.Commands
             {
                 var hotel = _context.Hotel.Where(a => a.Id == request.hotelId).FirstOrDefault();
                 if (hotel == null) return default;
-                var faciliteshotel = _context.FacilitesHotel.Where(a => a.hotelId == request.hotelId);
-               
+                if (hotel.Facilities!=null)
+                {
+                    var faciliteshotel = _context.FacilitesHotel.Where(a => a.hotelId == request.hotelId);
+
                     foreach (var item in faciliteshotel)
                     {
                         _context.FacilitesHotel.Remove(item);
                     }
+
+                }
+               
 
                 _context.Hotel.Remove(hotel);
                 await _context.SaveChangesAsync();
